Integrate github-token relevant tests into main run
per review comments https://github.com/w3c/browser-specs/pull/1023#pullrequestreview-1556230393
