+++ .github/workflows/lint.yml
@@ -24,18 +24,12 @@ jobs:
       - name: Install dependencies
         run: npm ci
 
-      - name: Test (including GitHub API tests)
+      - name: Test
         env:
           GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         if: ${{ env.GH_TOKEN }}
         run: |
           npm run test
-
-      - name: Test (without GitHub API tests)
-        env:
-          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        if: ${{ !env.W3C_API_KEY }}
-        run: npm run test-pr
 
       - name: Lint
         run: npm run lint
